isolate(dispatchers): do not send messages to closed channels

+++ isolate/spawndispatch.go
@@ -2,6 +2,7 @@ package isolate
 
 import (
 	"fmt"
+	"sync/atomic"
 
 	"golang.org/x/net/context"
 )
@@ -15,12 +16,14 @@ const (
 
 type spawnDispatch struct {
 	ctx     context.Context
+	killed  *uint32
 	process <-chan Process
 }
 
-func newSpawnDispatch(ctx context.Context, prCh <-chan Process) *spawnDispatch {
+func newSpawnDispatch(ctx context.Context, prCh <-chan Process, flagKilled *uint32) *spawnDispatch {
 	return &spawnDispatch{
 		ctx:     ctx,
+		killed:  flagKilled,
 		process: prCh,
 	}
 }
@@ -32,12 +35,14 @@ func (d *spawnDispatch) Handle(msg *message) (Dispatcher, error) {
 			select {
 			case pr, ok := <-d.process:
 				if ok {
-					if err := pr.Kill(); err != nil {
-						reply(d.ctx, replyKillError, errKillError, err.Error())
-						return
-					}
+					if atomic.CompareAndSwapUint32(d.killed, 0, 1) {
+						if err := pr.Kill(); err != nil {
+							reply(d.ctx, replyKillError, errKillError, err.Error())
+							return
+						}
 
-					reply(d.ctx, replyKillOk, nil)
+						reply(d.ctx, replyKillOk, nil)
+					}
 				}
 			case <-d.ctx.Done():
 			}
